// Web/RMA-Web-Shell/projects/debtcare/src/app/debt-manager/views/shared/models/age-analyisi-details.ts
export class AgeAnalysisDetails {
    policyId: number
    currentBalance: number
    balance30Days: number
    balance60Days: number
    balance90Days: number
    balance120Days: number
    netBalance: number
    id: number
    isActive: boolean
    isDeleted: boolean
    createdBy: any
    createdDate: string
    modifiedBy: any
    modifiedDate: string
  }